inndf:  report buffindexed usage with 2 decimals
Output was truncated at the inferior number (e.g. 37.00%).

+++ storage/overview.c
@@ -409,10 +409,10 @@ overview_config_set(struct overview *overview, struct overview_config *config)
 float
 overview_free_space(struct overview *overview)
 {
-    int space;
+    float space;
 
     if (overview->method->ctl(OVSPACE, &space))
         return space;
     else
-        return -1;
+        return -1.0f;
 }
